--- a/app/views/admin/registrations/_questions.html.haml
+++ b/app/views/admin/registrations/_questions.html.haml
@@ -1,9 +1,5 @@
-- @conference.questions.each do |q|
-
-  %b Q:
-  = q.title
-
-  %b A:
-  - registration.qanswers.where(question_id: q.id).each do |qa|
-    = qa.answer.title
-  %br
+- cache [:admin, @conference, registration.qanswers] do
+  %dl
+    - registration.qanswers.each do |qa|
+      %dt= qa.question.title
+      %dd= qa.answer.title