* Add some caching
* Reduce query count, including N+1s on questions & roles

**Checklist**

- [x] I have read the [Contribution & Best practices Guide](https://github.com/openSUSE/osem/blob/master/CONTRIBUTING.md).
- [x] My branch is up-to-date with the upstream `master` branch.
- [x] The tests pass locally with my changes.

**Short description of what this resolves/which [issues](https://github.com/openSUSE/osem/issues) does this fix?:**

<!-- List the issue number resolved with this change; if there is no open issue, describe the problem this request solves -->

- Registrations page was so slow it wouldn't load anymore (Total render time: 123 seconds)

**Changes proposed in this pull request:**

<!-- Summarize the changes, using declarative language. -->

- Add some caching
- Reduce query count, including N+1s on questions & roles

Total render time: 6 seconds. Still slow, but, you know, 1/20 of the time before ;-)