@bear454 2018-05-22
@Ana06 It's been > 3 weeks since last comments. Do you have any further issues, or should I consider your comments resolved?