@bear454 2018-09-19
@hennevogel I'm closing it; Moving to ajax-datatables will make most of the improvements moot, anyway. Thanks for the reminder.